Added field mapping
System will now map from the form field names to the target endpoint names
Added field and labels to the config page for the plugin
orgid and recordtype (along with the external flag) must be sent.
All other fields are sent if they are in the mapping list.
First in wins the field makes multiple mappings possible with the same settings.
Moved the curl post variables to an instance variable.